+++ packages/common/src/index.ts
@@ -52,5 +52,6 @@ export { default as ConcordiumGRPCClient } from './GRPCClient';
 export { getAccountTransactionHandler } from './accountTransactions';
 export { calculateEnergyCost } from './energyCost';
 
+export * from './uleb128';
 export * from './cis2';
 export * from './cis0';